First sketch: single trade contract
The contract allows 2 parties to negotiate a price for a single trade of assets X and Y.

Party A deploys a contract announcing that they want to trade certain amount of asset X (the amount being stored on chain).
They also commit to a price by storing a hash of the minimum amount of asset Y which they are willing to accept in exchange for the given amount of asset X.

Party B submits a transaction with a commitment to buy the given amount of asset X for some maximum amount of asset Y. Only the hash of Y amount is published on chain.

Party A and B then reveal the amounts of asset Y to which they have previously committed. If the prices cross, final the settlement price is calculated as the average of the to amounts of Y asset submitted by the parties.

next steps:
1. Add max time whiting which Y amounts need to be revealed, and punishment for not revealing in time.
2. Use off-chain storage to allow multiple orders and implement mutli-order batches
3. Use salt when creating commitment hashes